[ICC-232] 코드래빗 리뷰 반영

+++ modules/quiz/impl/src/main/java/com/icc/qasker/quiz/repository/ProblemRepository.java
@@ -12,6 +12,6 @@ public interface ProblemRepository extends JpaRepository<Problem, ProblemId> {
 
     Optional<Problem> findByIdProblemSetIdAndIdNumber(Long problemSetId, int number);
 
-    int countByIdProblemSetId(Long id);
+    long countByIdProblemSetId(Long id);
 }
